requirements.txt
# Gold Tier Autonomous AI Employee Dependencies

# Core
python-dotenv>=1.0.0
anthropic>=0.18.0

# Async HTTP
aiohttp>=3.9.0
aiofiles>=23.2.0

# Data validation
pydantic>=2.5.0

# Scheduling
apscheduler>=3.10.0

# Odoo JSON-RPC (handled by aiohttp)

# Social Media APIs
tweepy>=4.14.0  # Twitter/X

# Utilities
python-dateutil>=2.8.0
typing-extensions>=4.9.0

# Testing (optional)
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development (optional)
black>=23.12.0
isort>=5.13.0
mypy>=1.8.0
